Added checks for duplicate entries to UIs

+++ Gold_Badge_Challenges/Gold_Badge_Challenges/ProgramUI.cs
@@ -22,7 +22,7 @@ namespace C1_Cafe_Console
             while (loopMenuOptions)
             {
                 //Menu List
-                Console.WriteLine(" MENU OPTIONS:\n" +
+                Console.WriteLine("\n MENU OPTIONS:\n" +
                     " (1) Add new meal to the menu\n" +
                     " (2) Delete a meal from the menu\n" +
                     " (3) See all meals in the menu\n" +
@@ -67,16 +67,34 @@ namespace C1_Cafe_Console
             Console.Clear();
             CafeMenu newMenuItem = new CafeMenu();
 
-            Console.Write(" Enter number for the new meal: ");
-            newMenuItem.MealNumber = int.Parse(Console.ReadLine());           
+            bool alreadyExists = false;
+            while (!alreadyExists)
+            {
+                Console.Write(" Enter a number for the new meal (ex: 1, 2, 3 etc): ");
+                int checkMenuNumber = int.Parse(Console.ReadLine());
+                alreadyExists = CheckIfMealNumberAlreadyExists(checkMenuNumber);
+                if (alreadyExists == true)
+                {
+                    newMenuItem.MealNumber = checkMenuNumber;
+                }
+                else
+                {
+                    Console.WriteLine("\n This meal number already exists. Please choose another number.");
+                    Console.WriteLine("\n Press any key to continue...");
+                    Console.ReadKey();
+                    Console.Clear();
+                }
+            }
             
+            
+
             Console.Write(" Enter a name for the new meal: ");
             newMenuItem.MealName = Console.ReadLine();
 
             Console.Write(" Enter a description for the new meal: ");
             newMenuItem.MealDescription = Console.ReadLine();
 
-            Console.WriteLine(" Enter an ingredient then press enter. Type 'end' when complete");
+            Console.WriteLine(" Enter an ingredient then press enter. Type 'end' when complete.");
             bool input = true;
             while (input)
             {
@@ -93,7 +111,8 @@ namespace C1_Cafe_Console
             }
 
             Console.Write(" Enter the price for this meal: ");
-            newMenuItem.MealPrice = decimal.Parse(Console.ReadLine());
+            string amountAsString = Console.ReadLine();
+            newMenuItem.MealPrice = ConvertStringToDecimal(amountAsString);             
 
             bool wasAdded = _menuItemsRepo.AddMenuItemToList(newMenuItem);
             if (wasAdded)
@@ -112,20 +131,28 @@ namespace C1_Cafe_Console
             //Display all menu items
             SeeMenuItems();
 
-            //Get user input
-            Console.Write("\n Enter meal number to delete from menu list: ");
-            int input = int.Parse(Console.ReadLine());
-
-            //Call delete method
-            bool wasDeleted = _menuItemsRepo.DeleteMenuItemFromList(input);
-
-            if (wasDeleted)
+            List<CafeMenu> listOfMenuItems = _menuItemsRepo.GetMenuItemsList();
+            if (listOfMenuItems.Count != 0)
             {
-                Console.WriteLine("\n The meal was successfully deleted.");
+                //Get user input
+                Console.Write("\n Enter the meal number to delete from menu list: ");
+                int input = int.Parse(Console.ReadLine());
+
+                //Call delete method
+                bool wasDeleted = _menuItemsRepo.DeleteMenuItemFromList(input);
+
+                if (wasDeleted)
+                {
+                    Console.WriteLine("\n The meal was successfully deleted.");
+                }
+                else
+                {
+                    Console.WriteLine("\n The meal could not be deleted.");
+                }
             }
             else
             {
-                Console.WriteLine("\n The meal could not be deleted");
+                Console.WriteLine("\n There are no menu items to delete.");
             }
         }
 
@@ -134,29 +161,60 @@ namespace C1_Cafe_Console
             Console.Clear();
             List<CafeMenu> listOfMenuItems = _menuItemsRepo.GetMenuItemsList();
 
-            foreach (CafeMenu items in listOfMenuItems)
+            if (listOfMenuItems.Count != 0)
             {
-                Console.WriteLine($" Meal number: {items.MealNumber}\n" +
-                    $" Meal name: {items.MealName}\n" +
-                    $" Meal description: {items.MealDescription}");
-
-                //Console.WriteLine(" Ingredients: ");
-                Console.Write(" Ingredients: ");
-                int count = items.MealIngredients.Count;
-                foreach (string ingredient in items.MealIngredients)
+                foreach (CafeMenu items in listOfMenuItems)
                 {
-                    if (count > 1)
+                    Console.WriteLine($" Meal number: {items.MealNumber}\n" +
+                        $" Meal name: {items.MealName}\n" +
+                        $" Meal description: {items.MealDescription}");
+
+                    //Console.WriteLine(" Ingredients: ");
+                    Console.Write(" Ingredients: ");
+                    int count = items.MealIngredients.Count;
+                    foreach (string ingredient in items.MealIngredients)
                     {
-                        Console.Write($"{ingredient}, ");
-                        count--;
+                        if (count > 1)
+                        {
+                            Console.Write($"{ingredient}, ");
+                            count--;
+                        }
+                        else
+                        {
+                            Console.Write($"{ingredient}\n");
+                        }
                     }
-                    else
-                    {
-                        Console.Write($"{ingredient}\n");
-                    }                    
+                    //Console.WriteLine($" Meal Price: ${items.MealPrice}\n");
+                    Console.WriteLine(" Meal Price: ${0:F2}\n", items.MealPrice);
                 }
-                //Console.WriteLine($" Meal Price: ${items.MealPrice}\n");
-                Console.WriteLine(" Meal Price: ${0:F2}\n", items.MealPrice);
+            }
+            else
+            {
+                Console.WriteLine("\n There are no menu items to display.");
+            }
+        }
+
+        private decimal ConvertStringToDecimal(string valueAsString)
+        {
+            if (valueAsString.StartsWith("$"))
+            {
+                valueAsString = valueAsString.Substring(1);
+            }
+
+            decimal valueAsDecimal = decimal.Parse(valueAsString);
+            return valueAsDecimal;
+        }
+
+        private bool CheckIfMealNumberAlreadyExists(int valueAsInt)
+        {
+            CafeMenu checkMealNumber = _menuItemsRepo.GetMenuItemByNumber(valueAsInt);
+            if (checkMealNumber == null)
+            {
+                return true;
+            }
+            else
+            {
+                return false;
             }
         }
     }
